Correct float pretty-printing
MLton's Real64.toString prints doubles that have no fractional part the
same as integers. This doesn't play well with C, since C won't convert
integers to doubles unless it absolutely has to; for example, it will
evaluate 3/5 to 0 instead of 0.6. To avoid this, always print doubles in
scientific notation, which forces the C parser to parse them as doubles.